feat(ToyRobot): introduce new ToyRobot data class
This commit introduces our new ToyRobot class that will be responsible
for holding our current Position and Direction. It also will know how
to place itself onto a given table. This commit also adds Automaton,
a class that will be responsible for dealing with movement actions that
will depend on its cardinal direction.